fix: move all tests to one file
